+++ packages/wrangler/src/dev/inspect.ts
@@ -8,11 +8,12 @@ import { URL, fileURLToPath, pathToFileURL } from "node:url";
 import open from "open";
 import { useEffect, useRef, useState } from "react";
 import WebSocket, { WebSocketServer } from "ws";
-import { version } from "../package.json";
-import { logger } from "./logger";
+import { version } from "../../package.json";
+import { logger } from "../logger";
+import { getSourceMappedStack } from "../sourcemap";
+import { getAccessToken } from "../user/access";
 import { waitForPortToBeAvailable } from "./proxy";
-import { getSourceMappedStack } from "./sourcemap";
-import { getAccessToken } from "./user/access";
+import type { SourceMapMetadata } from "../deployment-bundle/bundle";
 import type Protocol from "devtools-protocol";
 import type { IncomingMessage, Server, ServerResponse } from "node:http";
 import type { RawSourceMap } from "source-map";
@@ -35,19 +36,10 @@ import type { MessageEvent } from "ws";
  * and exceptions)
  */
 
-/**
- * TODO:
- * - clear devtools whenever we save changes to the worker
- * - clear devtools when we switch between local/remote modes
- * - handle more methods from console
- */
-
-// Information about Wrangler's bundling process that needs passsed through
-// for DevTools sourcemap transformation
-export interface SourceMapMetadata {
-	tmpDir: string;
-	entryDirectory: string;
-}
+// TODO:
+// - clear devtools whenever we save changes to the worker
+// - clear devtools when we switch between local/remote modes
+// - handle more methods from console
 
 interface InspectorProps {
 	/**
@@ -618,14 +610,13 @@ export default function useInspector(props: InspectorProps) {
 }
 
 /**
- * This function converts a message serialised as a devtools event
+ * This function converts a message serialized as a devtools event
  * into arguments suitable to be called by a console method, and
  * then actually calls the method with those arguments. Effectively,
  * we're just doing a little bit of the work of the devtools console,
  * directly in the terminal.
  */
-
-export const mapConsoleAPIMessageTypeToConsoleMethod: {
+const mapConsoleAPIMessageTypeToConsoleMethod: {
 	[key in Protocol.Runtime.ConsoleAPICalledEvent["type"]]: Exclude<
 		keyof Console,
 		"Console"
