+++ packages/wrangler/src/dev/local.tsx
@@ -4,8 +4,8 @@ import { useEffect, useRef, useState } from "react";
 import onExit from "signal-exit";
 import { fetch } from "undici";
 import { registerWorker } from "../dev-registry";
-import useInspector from "../inspect";
 import { logger } from "../logger";
+import useInspector from "./inspect";
 import { MiniflareServer } from "./miniflare";
 import type { Config } from "../config";
 import type { CfWorkerInit, CfScriptFormat } from "../deployment-bundle/worker";
